GH-1723 Refactor RHS run list
